requirements.txt
qwen-agent[rag,code_interpreter,mcp]
torch
transformers==4.55.0
accelerate
datasets
huggingface-hub
sentence-transformers
numpy
soundfile
python-dotenv
pyyaml
fastapi
uvicorn
openai
google-genai
mistralai
requests
vllm[audio]==0.10.1.1
faster-whisper
